+++ src/core/core.ts
@@ -5,6 +5,7 @@ export * from "./env.js";
 export * from "./event.js";
 export * from "./isolate.js";
 export * from "./lifecycle.js";
+export * from "./overwrite.js";
 export * from "./signals.js";
 export type * from "./types.js";
 export * from "./unique-id.js";